refactor(desktop): split session and chat composables
Group desktop launcher, SDK steering, and chat-view data refactors behind smaller composables while preserving UI behavior.
