+++ apps/desktop/src/composables/useSessionLauncher.ts
@@ -1,16 +1,6 @@
-import type { LaunchConfig, SessionTemplate } from "@tracepilot/types";
-import { DEFAULT_CLI_COMMAND, DEFAULT_MODEL_ID, getTierLabel } from "@tracepilot/types";
-import { formatCost, useClipboard, useConfirmDialog, useToast } from "@tracepilot/ui";
-import {
-  computed,
-  type InjectionKey,
-  inject,
-  onMounted,
-  onUnmounted,
-  reactive,
-  ref,
-  watch,
-} from "vue";
+import { DEFAULT_MODEL_ID } from "@tracepilot/types";
+import { formatCost, useClipboard, useToast } from "@tracepilot/ui";
+import { computed, type InjectionKey, inject, onMounted, onUnmounted, ref, watch } from "vue";
 import { useRoute, useRouter } from "vue-router";
 import { browseForDirectory } from "@/composables/useBrowseDirectory";
 import { useGitRepository } from "@/composables/useGitRepository";
@@ -20,13 +10,23 @@ import { usePreferencesStore } from "@/stores/preferences";
 import { useSdkStore } from "@/stores/sdk";
 import { useSessionsStore } from "@/stores/sessions";
 import { useWorktreesStore } from "@/stores/worktrees";
+import { useLauncherCliPreview } from "./sessionLauncher/useLauncherCliPreview";
+import { useLauncherForm } from "./sessionLauncher/useLauncherForm";
+import { useLauncherTemplates } from "./sessionLauncher/useLauncherTemplates";
+
+export type { CliPart, ReasoningEffort } from "./sessionLauncher/types";
 
 /**
  * Central state + action coordinator for `SessionLauncherView`.
  *
- * Owns the launch form state, saved-template lifecycle, git repository
- * operations, environment variable rows, context-menu state, and the CLI
- * command preview derivations. Shared with child components via
+ * Internally composed from three focused composables:
+ *   - `useLauncherForm` — reactive form state, validation, launchConfig
+ *   - `useLauncherTemplates` — template selection, lifecycle, context menu
+ *   - `useLauncherCliPreview` — derived CLI preview (single source via
+ *     `buildLaunchCliArgs`)
+ *
+ * The flat exported API is preserved for backwards compatibility with
+ * `SessionLauncherView` and child components that consume it via
  * `provide`/`inject` (`SessionLauncherKey` + `useSessionLauncherContext`).
  */
 export function useSessionLauncher() {
@@ -37,45 +37,58 @@ export function useSessionLauncher() {
   const router = useRouter();
   const launching = ref(false);
   const { success: toastSuccess, error: toastError } = useToast();
-  const { confirm } = useConfirmDialog();
   const { copy: copyToClipboard } = useClipboard();
 
-  const repoPath = ref("");
-  const branch = ref("");
-  const selectedModel = ref("");
-  const createWorktree = ref(false);
-  const autoApprove = ref(false);
-  const headless = ref(false);
-  const uiServer = ref(false);
-  const reasoningEffort = ref<"low" | "medium" | "high">("medium");
-  const prompt = ref("");
-  const customInstructions = ref("");
-  const envVars = reactive<{ key: string; value: string }[]>([]);
+  const sdkFeatureEnabled = computed(() => prefsStore.isFeatureEnabled("copilotSdk"));
+  const storeLoading = computed(() => store.loading);
 
-  const selectedTemplateId = ref<string | null>(null);
-  const showAdvanced = ref(false);
-  const showTemplateForm = ref(false);
-  const templateForm = reactive({ name: "", description: "", category: "", icon: "" });
-  const contextMenuTpl = ref<{ id: string; x: number; y: number } | null>(null);
-  const confirmingDeleteId = ref<string | null>(null);
+  const form = useLauncherForm({ sdkFeatureEnabled, loading: storeLoading });
+  const {
+    repoPath,
+    branch,
+    baseBranch,
+    selectedModel,
+    createWorktree,
+    autoApprove,
+    headless,
+    uiServer,
+    reasoningEffort,
+    prompt,
+    customInstructions,
+    envVars,
+    launchConfig,
+    canLaunch,
+    addEnvVar,
+    removeEnvVar,
+  } = form;
 
-  const baseBranch = ref("");
+  const templates = useLauncherTemplates({ form, launchConfig });
+  const {
+    selectedTemplateId,
+    showAdvanced,
+    showTemplateForm,
+    templateForm,
+    contextMenuTpl,
+    confirmingDeleteId,
+    selectedTemplateName,
+    hasDismissedDefaults,
+    tierLabel,
+    templateIcon,
+    templateDisplayName,
+    applyTemplate,
+    clearTemplateSelection,
+    moveTemplate,
+    deleteTemplateInline,
+    cancelDeleteInline,
+    openContextMenu,
+    deleteContextTemplate,
+    closeContextMenu,
+    handleSaveTemplate,
+  } = templates;
+
+  const { cliCommand, cliCommandParts } = useLauncherCliPreview(launchConfig);
 
   const selectedModelInfo = computed(() => store.models.find((m) => m.id === selectedModel.value));
-  const sdkFeatureEnabled = computed(() => prefsStore.isFeatureEnabled("copilotSdk"));
-
-  const selectedTemplateName = computed(() => {
-    if (!selectedTemplateId.value) return "Custom";
-    return store.templates.find((t) => t.id === selectedTemplateId.value)?.name ?? "Custom";
-  });
-
-  const envVarsRecord = computed(() => {
-    const rec: Record<string, string> = {};
-    for (const e of envVars) {
-      if (e.key.trim()) rec[e.key.trim()] = e.value;
-    }
-    return rec;
-  });
 
   const {
     defaultBranch,
@@ -102,61 +115,6 @@ export function useSessionLauncher() {
     clearTemplateSelection();
   }
 
-  const launchConfig = computed<LaunchConfig>(() => ({
-    repoPath: repoPath.value,
-    branch: branch.value || undefined,
-    baseBranch: createWorktree.value && baseBranch.value ? baseBranch.value : undefined,
-    model: selectedModel.value || undefined,
-    prompt: prompt.value || undefined,
-    customInstructions: customInstructions.value || undefined,
-    reasoningEffort: reasoningEffort.value,
-    headless: headless.value,
-    createWorktree: createWorktree.value,
-    autoApprove: autoApprove.value,
-    uiServer: !headless.value && uiServer.value,
-    launchMode: headless.value ? "sdk" : "terminal",
-    envVars: envVarsRecord.value,
-    cliCommand: prefsStore.cliCommand || DEFAULT_CLI_COMMAND,
-  }));
-
-  const effectiveCli = computed(() => prefsStore.cliCommand || DEFAULT_CLI_COMMAND);
-
-  const cliCommand = computed(() => {
-    if (launchConfig.value.launchMode === "sdk") {
-      return "Copilot SDK bridge (headless session)";
-    }
-    const parts = [effectiveCli.value];
-    if (launchConfig.value.model) parts.push(`--model ${launchConfig.value.model}`);
-    if (launchConfig.value.autoApprove) parts.push("--allow-all");
-    if (launchConfig.value.uiServer) parts.push("--ui-server");
-    if (launchConfig.value.reasoningEffort)
-      parts.push(`--reasoning-effort ${launchConfig.value.reasoningEffort}`);
-    if (launchConfig.value.prompt) {
-      parts.push(`--interactive '${launchConfig.value.prompt.replace(/'/g, "''")}'`);
-    }
-    return parts.join(" ");
-  });
-
-  const cliCommandParts = computed<{ flag: string; value?: string }[]>(() => {
-    if (launchConfig.value.launchMode === "sdk") {
-      return [{ flag: "Copilot SDK bridge" }, { flag: "headless session" }];
-    }
-    const parts: { flag: string; value?: string }[] = [{ flag: effectiveCli.value }];
-    if (launchConfig.value.model) {
-      parts.push({ flag: "--model", value: launchConfig.value.model });
-    }
-    if (launchConfig.value.autoApprove) parts.push({ flag: "--allow-all" });
-    if (launchConfig.value.uiServer) parts.push({ flag: "--ui-server" });
-    if (launchConfig.value.reasoningEffort) {
-      parts.push({ flag: "--reasoning-effort", value: launchConfig.value.reasoningEffort });
-    }
-    if (launchConfig.value.prompt) {
-      parts.push({ flag: "--interactive", value: launchConfig.value.prompt });
-    }
-    return parts;
-  });
-
-  // Preview path for worktree creation
   const worktreePreviewPath = computed(() => {
     if (!createWorktree.value || !branch.value) return "";
     return computeWorktreePath(branch.value).replace(/\//g, "\\");
@@ -169,94 +127,6 @@ export function useSessionLauncher() {
     if (pr === 0) return "Free";
     return `~${formatCost(cost)} (${pr}x premium requests)`;
   });
-
-  const canLaunch = computed(() => {
-    if (!repoPath.value.trim() || store.loading) return false;
-    if (createWorktree.value && !branch.value.trim()) return false;
-    if (launchConfig.value.launchMode === "sdk" && !sdkFeatureEnabled.value) return false;
-    return true;
-  });
-
-  const defaultTemplateIds = ["default-multi-agent-review", "default-write-tests"];
-  const hasDismissedDefaults = computed(() =>
-    defaultTemplateIds.some((id) => !store.templates.some((t) => t.id === id)),
-  );
-
-  function tierLabel(tier: string): string {
-    if (tier === "premium" || tier === "standard" || tier === "fast") {
-      return getTierLabel(tier);
-    }
-    // For non-tier strings (e.g. reasoning effort: low/medium/high), capitalize
-    return tier.charAt(0).toUpperCase() + tier.slice(1);
-  }
-
-  function extractEmoji(name: string): string {
-    const match = name.match(/^(\p{Emoji_Presentation}|\p{Extended_Pictographic})/u);
-    return match ? match[0] : "📄";
-  }
-
-  function templateIcon(tpl: SessionTemplate): string {
-    return tpl.icon || extractEmoji(tpl.name);
-  }
-
-  function templateDisplayName(name: string): string {
-    return name.replace(/^(\p{Emoji_Presentation}|\p{Extended_Pictographic})\s*/u, "");
-  }
-
-  function applyTemplate(tplId: string) {
-    if (selectedTemplateId.value === tplId) {
-      selectedTemplateId.value = null;
-      return;
-    }
-    const tpl = store.templates.find((t: SessionTemplate) => t.id === tplId);
-    if (!tpl) return;
-    selectedTemplateId.value = tplId;
-    if (tpl.config.repoPath) {
-      repoPath.value = tpl.config.repoPath;
-    }
-    branch.value = tpl.config.branch ?? "";
-    selectedModel.value = tpl.config.model ?? "";
-    createWorktree.value = tpl.config.createWorktree;
-    baseBranch.value = tpl.config.baseBranch ?? "";
-    autoApprove.value = tpl.config.autoApprove;
-    headless.value = tpl.config.headless;
-    uiServer.value = tpl.config.uiServer ?? false;
-    reasoningEffort.value = (tpl.config.reasoningEffort as "low" | "medium" | "high") ?? "medium";
-    prompt.value = tpl.config.prompt ?? "";
-    customInstructions.value = tpl.config.customInstructions ?? "";
-    envVars.length = 0;
-    if (tpl.config.envVars) {
-      for (const [k, v] of Object.entries(tpl.config.envVars)) {
-        envVars.push({ key: k, value: v });
-      }
-    }
-  }
-
-  function clearTemplateSelection() {
-    selectedTemplateId.value = null;
-  }
-
-  function moveTemplate(idx: number, direction: "up" | "down") {
-    const target = direction === "up" ? idx - 1 : idx + 1;
-    if (target < 0 || target >= store.templates.length) return;
-    const arr = [...store.templates];
-    [arr[idx], arr[target]] = [arr[target], arr[idx]];
-    store.templates = arr;
-  }
-
-  async function deleteTemplateInline(tplId: string) {
-    if (confirmingDeleteId.value === tplId) {
-      await store.deleteTemplate(tplId);
-      if (selectedTemplateId.value === tplId) selectedTemplateId.value = null;
-      confirmingDeleteId.value = null;
-    } else {
-      confirmingDeleteId.value = tplId;
-    }
-  }
-
-  function cancelDeleteInline() {
-    confirmingDeleteId.value = null;
-  }
 
   function selectRecentRepo(event: Event) {
     const val = (event.target as HTMLSelectElement).value;
@@ -272,14 +142,6 @@ export function useSessionLauncher() {
       repoPath.value = dir;
       clearTemplateSelection();
     }
-  }
-
-  function addEnvVar() {
-    envVars.push({ key: "", value: "" });
-  }
-
-  function removeEnvVar(idx: number) {
-    envVars.splice(idx, 1);
   }
 
   async function handleLaunch() {
@@ -323,56 +185,6 @@ export function useSessionLauncher() {
     }
   }
 
-  async function handleSaveTemplate() {
-    if (!templateForm.name.trim()) return;
-    const existing = store.templates.find(
-      (t) => t.name.toLowerCase() === templateForm.name.trim().toLowerCase(),
-    );
-    if (existing) {
-      const { confirmed } = await confirm({
-        title: "Overwrite Template",
-        message: `Template '${existing.name}' already exists. Do you want to overwrite it?`,
-        variant: "warning",
-        confirmLabel: "Overwrite",
-      });
-      if (!confirmed) return;
-    }
-    await store.saveTemplate({
-      id: existing?.id ?? crypto.randomUUID(),
-      name: templateForm.name,
-      description: templateForm.description,
-      category: templateForm.category,
-      icon: templateForm.icon.trim() || undefined,
-      tags: [],
-      config: launchConfig.value,
-      createdAt: existing?.createdAt ?? new Date().toISOString(),
-      usageCount: existing?.usageCount ?? 0,
-    });
-    showTemplateForm.value = false;
-    templateForm.name = "";
-    templateForm.description = "";
-    templateForm.category = "";
-    templateForm.icon = "";
-  }
-
-  function openContextMenu(e: MouseEvent, tplId: string) {
-    e.preventDefault();
-    contextMenuTpl.value = { id: tplId, x: e.clientX, y: e.clientY };
-  }
-
-  async function deleteContextTemplate() {
-    if (!contextMenuTpl.value) return;
-    await store.deleteTemplate(contextMenuTpl.value.id);
-    if (selectedTemplateId.value === contextMenuTpl.value.id) {
-      selectedTemplateId.value = null;
-    }
-    contextMenuTpl.value = null;
-  }
-
-  function closeContextMenu() {
-    contextMenuTpl.value = null;
-  }
-
   async function copyCommand() {
     const ok = await copyToClipboard(cliCommand.value);
     if (ok) toastSuccess("Command copied to clipboard");
@@ -391,7 +203,6 @@ export function useSessionLauncher() {
     store.initialize();
     document.addEventListener("click", closeContextMenu);
 
-    // Load registered repos for the dropdown, discovering from sessions if needed
     if (worktreeStore.registeredRepos.length === 0) {
       await worktreeStore.loadRegisteredRepos();
       if (worktreeStore.registeredRepos.length === 0) {
@@ -399,7 +210,6 @@ export function useSessionLauncher() {
       }
     }
 
-    // Pre-fill from query params (e.g., navigated from Worktree Manager)
     if (route.query.repoPath) {
       repoPath.value = String(route.query.repoPath);
       worktreeStore.loadBranches(repoPath.value);
